requirements.txt
torch
torchvision
numpy
matplotlib
scikit-learn
tqdm
seaborn
flask
Pillow
